Starting blocks
I have found an API that calculates the distance between two destinations and depending on the type of transport, it calculates how much C02 emissions is released. I have set the request parameters to allow the API to be read and parse JSON response so when the API calculates the co2 emissions and distance, I can extract the information  and print it. Next thing I have to do is add more types of transport which i may do by making my own API or even just using the ones provided for me from the other API.