Rest API created
I have created my rest api but i am having issues with springboot running so I have to look into that. I am also deciding whether it is easier to get rid of the EmissionsCalculatorUtils.java as i want to combine the two api together but it is proving difficult for me to code the it to allow for an extra transportation method.