complete code change
I created three endpoints, one for the trees, destination and third party emissions calculator.  I also created two json files that hold trees and the fujitsu offices. When ran, it will save in a new file with all information combined. Having issues with adding and deleting in CRUD due to .UnsupportedOperationException error. Stuggling to fix this