Progress
I refactored the code so that the destination address for the third part API could be reached from it ID which i used in my emissions calculator. I also formatted the JSON files, changed my code so the abstractjourney could hold the parent classes to the code while journeyrequest and emissionsdata used it as child classes. I added code to quicksort the journeys saved in the JSON into co2e order.