changes
i fixed some POM issues and deleted the prettier npm dev dependencies in my node_modules.  I removed  have add, update, delete in the path of my endpoints as the request type determines the operation. I didnt want to allow the id for an address to be changed, so i added a constructor that accepts an id in destinationaddress and emissionsdata.  i then created a new instance to pass it.